Test interval arrays only for `arrow>=52`
